Moved CSS parsing from traits.rs to css.rs, updated docs
